refactored my code and rearranged component files
