+++ starwars/src/App.js
@@ -1,28 +1,24 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
+
 import './App.css';
-import CharactersContent from './components/CharacterName/CharactersContent';
+import CharactersContent from './components/Characters/CharactersContent';
 
 const App = () => {
-  // Try to think through what state you'll need for this app before starting. Then build out
-  // the state properties here.
-
-  // Fetch characters from the star wars api in an effect hook. Remember, anytime you have a 
-  // side effect in a component, you want to think about which state and/or props it should
-  // sync up with, if any.
+  //Creating state for the pulled data
   const [characters, setCharacters] = useState([]);
-
+  const [isLoading, setLoading] = useState(true);
+  //Passing the data to state
   useEffect(() => {
         axios.get("https://swapi.co/api/people/")
         .then((response) => {
-          console.log(response.data.results);
           setCharacters(response.data.results);
         })
         .catch((error) => {
            return error;
         })   
   },[])
-
+  //rendering the components
   return (
     <div className="App">
       <h1 className="Header">React Wars</h1>
@@ -31,11 +27,9 @@ const App = () => {
           return (
             <CharactersContent key={index} characters={character}/>
           )
-
         })
     } 
     </div>
   );
 }
-
 export default App;
