+++ starwars/src/components/Characters/CharHeight/CharacterHeight.js
@@ -1,21 +1,14 @@
 import React from "react";
 import styled from "styled-components";
 
-//import CharName from "./CharName";
-
 function CharactersHeight (props) {
-    console.log(props);
-
+    //styling the elements
     const CharHeightStyles = styled.p`
     font-family: 'Courier New', Courier, monospace;
     `;
-    //const {characters} = props;
-
     return (
         <>
-            {
-                <CharHeightStyles>Height:{props.charheight} cm</CharHeightStyles>
-            }
+            <CharHeightStyles>Height:{props.charheight} cm</CharHeightStyles>
         </>
     )
 }
