+++ starwars/src/components/Characters/CharName/CharName.js
@@ -1,6 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
+//styling the elements
 const CharacterNameStyles = styled.p`
 font-family: 'Courier New', Courier, monospace;
 `;
@@ -10,8 +11,8 @@ function CharName (props) {
 
     return (
         <>
-        <CharacterNameStyles>Name:{props.charname}</CharacterNameStyles>
-    </>
+            <CharacterNameStyles>Name:{props.charname}</CharacterNameStyles>
+        </>
     )
 
 }
